+++ data-sketches-core/src/DataSketches/Quantiles/RelativeErrorQuantile/Types.hs
@@ -1,25 +1,18 @@
-module DataSketches.Quantiles.RelativeErrorQuantile.Types where
-import Control.Monad.Primitive
-import qualified DataSketches.Quantiles.RelativeErrorQuantile.Internal.InequalitySearch as IS
+module DataSketches.Quantiles.RelativeErrorQuantile.Types
+  ( RankAccuracy(..)
+  , DoubleIsNonFiniteException(..)
+  ) where
 
-data Criterion = (:<) | (:<=)
-  deriving (Show, Eq)
+import Control.Exception (Exception)
 
-instance IS.InequalitySearch Criterion where
-  inequalityCompare c = case c of
-    (:<) -> IS.inequalityCompare (IS.:<)
-    (:<=) -> IS.inequalityCompare(IS.:<=)
-  resolve c = case c of
-    (:<) -> IS.resolve (IS.:<)
-    (:<=) -> IS.resolve (IS.:<=)
-  getIndex c = case c of
-    (:<) -> IS.getIndex (IS.:<)
-    (:<=) -> IS.getIndex (IS.:<=)
-
-
-data RankAccuracy 
-  = HighRanksAreAccurate 
+data RankAccuracy
+  = HighRanksAreAccurate
   -- ^ High ranks are prioritized for better accuracy.
   | LowRanksAreAccurate
   -- ^ Low ranks are prioritized for better accuracy
   deriving (Show, Eq)
+
+newtype DoubleIsNonFiniteException = DoubleIsNonFiniteException Double
+  deriving (Show, Eq)
+
+instance Exception DoubleIsNonFiniteException